+++ trixnity-client/src/commonMain/kotlin/net/folivo/trixnity/client/store/MediaCacheMapping.kt
@@ -7,6 +7,6 @@ import kotlinx.serialization.Serializable
 data class MediaCacheMapping(
     val cacheUri: String,
     val mxcUri: String? = null,
-    val size: Long? = null,
+    val size: Long = 0,
     val contentType: String? = ContentType.Application.OctetStream.toString(),
 )