+++ trixnity-client/trixnity-client-repository-realm/src/commonMain/kotlin/net/folivo/trixnity/client/store/repository/realm/RealmMediaCacheMappingRepository.kt
@@ -13,7 +13,7 @@ internal class RealmMediaCacheMapping : RealmObject {
     @PrimaryKey
     var cacheUri: String = ""
     var mxcUri: String? = null
-    var size: Long? = null
+    var size: Long = 0
     var contentType: String? = null
 }