+++ trixnity-client/trixnity-client-repository-room/src/commonMain/kotlin/net/folivo/trixnity/client/store/repository/room/TrixnityRoomDatabase.kt
@@ -7,7 +7,7 @@ import androidx.room.TypeConverters
 @Database(
     entities = [
         RoomAccount::class,
-        RoomServerVersions::class,
+        RoomServerData::class,
         RoomCrossSigningKeys::class,
         RoomDeviceKeys::class,
         RoomGlobalAccountData::class,
@@ -46,7 +46,7 @@ import androidx.room.TypeConverters
 )
 abstract class TrixnityRoomDatabase : RoomDatabase() {
     abstract fun account(): AccountDao
-    abstract fun serverVersions(): ServerVersionsDao
+    abstract fun serverData(): ServerDataDao
     abstract fun crossSigningKeys(): CrossSigningKeysDao
     abstract fun deviceKeys(): DeviceKeysDao
     abstract fun globalAccountData(): GlobalAccountDataDao