+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add utils for converting between Input-/OutputStream and ByteArrayFlow
 - Add extensions for creating ByteArrayFlow from ByteBuffer
+- Check server media config before uploading any media.
 
 ### Changed